Implement changes specified by the teacher
Improved code style
Deleted unnecessary variables
Reduced the usage of the surplus functions

+++ src/main/java/ru/mail/polis/homework/simple/IntegerTask.java
@@ -14,12 +14,12 @@ public class IntegerTask {
      * Пример: (5) -> 15
      */
     public static int sum(int n) {
-        if(n < 1){
+        if (n < 1) {
             return 0;
         }
 
         int sum = 0;
-        for(int i = 1; i <= n; i++){
+        for (int i = 1; i <= n; i++) {
             sum += i;
         }
         return sum;
@@ -32,23 +32,23 @@ public class IntegerTask {
      * Пример: (10, 3, 2) -> 8
      */
     public static int snake(int height, int top, int bottom) {
-        if(height <= 0 || top <= 0 || bottom <= 0){
+        if (height <= 0 || top <= 0 || bottom <= 0) {
             throw new IllegalArgumentException("Incorrect values");
         }
 
-        if(top >= height){
+        if (top >= height) {
             return 1;
         }
 
         int shift = top - bottom;
-        if(shift <= 0){
+        if (shift <= 0) {
             return Integer.MAX_VALUE;
         }
         byte extraDay = 0;
-        if((height - top)%shift != 0){
+        if ((height - top) % shift != 0) {
             extraDay = 1;
         }
-        return (height - top)/shift + 1 + extraDay;
+        return (height - top) / shift + 1 + extraDay;
     }
 
     /**
@@ -56,11 +56,11 @@ public class IntegerTask {
      * Пример: (454355, 3) -> 3
      */
     public static int kDecimal(int n, int order) {
-        if(order <= 0){
+        if (order <= 0) {
             throw new IllegalArgumentException("Invalid oreder");
         }
         int orderGetter = (int) Math.pow(10, order - 1);
-        if(Math.abs(n) < orderGetter){
+        if (Math.abs(n) < orderGetter) {
             throw new IllegalArgumentException("There is no such order in the number provided");
         }
 
@@ -73,12 +73,12 @@ public class IntegerTask {
      * Пример: (5) -> 120
      */
     public static long factorial(byte n) {
-        if(n < 0){
+        if (n < 0) {
             throw new IllegalArgumentException("Can't get a factorial of a negative number");
         }
 
         long factorial = 1;
-        for(int i = 1; i <= n; i++){
+        for (byte i = 1; i <= n; i++) {
             factorial *= i;
         }
